refactor: RecordFilters 컴포넌트와 스타일을 기능별로 완전 분리
