+++ src/components/memory/RecordFilters/RecordFilters.styled.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 16px;
+`;
